+++ src/global.d.ts
@@ -13,9 +13,9 @@ declare global {
 		// and what to do when importing types
 
 		interface Locals {
-			sessionToken?: string;
-			userID?: string;
-			state?: string;
+			// sessionToken?: string;
+			// userID?: string;
+			// state?: string;
 		}
 
 		// interface Platform {}