+++ src/lib/server/services/Account.ts
@@ -13,18 +13,23 @@ export type DBAccount = {
 	sessionTokens?: Set<string>;
 	apiKeys?: Record<string, Set<string>>;
 };
-export const UserAuth = z.object({
-	sessionToken: z.string(),
-	userID: z.string(),
-});
-export type UserAuth = z.infer<typeof UserAuth>;
 
-export const OAuth = z.object({
-	apiKey: z.string(),
-});
-export type OAuth = z.infer<typeof OAuth>;
+export const AuthToken = z.string();
+export type AuthToken = z.infer<typeof AuthToken>;
 
-export const Auth = z.union([OAuth, UserAuth]);
+export const ApiKey = z.string();
+export type ApiKey = z.infer<typeof ApiKey>;
+
+export const Auth = z.union([
+	z.object({
+		type: z.literal("AuthToken"),
+		authToken: AuthToken,
+	}),
+	z.object({
+		type: z.literal("ApiKey"),
+		apiKey: ApiKey,
+	}),
+]);
 export type Auth = z.infer<typeof Auth>;
 
 export type Edits = Pick<Partial<DBAccount>, "picture" | "bio" | "name" | "username">;